Add nginx has a HTTPS proxy on localhost:9193
This is really useful if you upload data from an other location.
NOTE:
  Update the file etc/appcfgd-nginx.conf to enable nginx to listen
  for example 0.0.0.0.

The buildout generate certificate and key for the SSL connection
The problem the certification. + appcfg use a verficiation of the certificate
by the SSL module.

For the moment I'm planning to path appcfg to remove this verification.